rename: 대소문자 관련 폴더명 변경 (#29)
