+++ src/components/composites/book/detail/index.ts
@@ -1,3 +1,3 @@
-export * from './Info';
+export * from './info';
 export * from './record';
 export { default as BookDetail } from './BookDetail';
