+++ src/components/composites/book/detail/info/BookDetailInfo.skeleton.tsx
